% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2_oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A nbaPlayers Function}
\usage{
oldest_player(data, year)
}
\arguments{
\item{data}{The input dataset}

\item{year}{The input value of year}
}
\description{
This function allows you to know the oldest player in the NBA data for that year
}
\examples{
oldest_player()
}
\keyword{year}
